Add connect sibling node function
Why: To provide an easy way to connect sibling ndoes witha  directional
edge whenever a new child is created

How: I created the connectSiblingNode function in cmdConverter.js. I
also added a function call to that function in the renderSplitNode
function.

In the future I hope to add the connectSiblingNode function to the
create child animation function so that whenver a new child is created
it automatically is connected to its siblings.